[refactor]: restructure platform integrations (bots → integrations)
